Factor out potentially problematic Word as usize casts

+++ fuel-vm/src/interpreter/alu/wideint.rs
@@ -61,11 +61,11 @@ macro_rules! wideint_ops {
                     let dest: &mut Word = &mut w[ra.try_into()?];
 
                     // LHS argument is always indirect, load it
-                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b as usize)?);
+                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b)?);
 
                     // RHS is only indirect if the flag is set
                     let rhs: $t = if args.indirect_rhs {
-                        $t::from_be_bytes(self.memory.read_bytes(c as usize)?)
+                        $t::from_be_bytes(self.memory.read_bytes(c)?)
                     } else {
                         c.into()
                     };
@@ -86,11 +86,11 @@ macro_rules! wideint_ops {
                     let (SystemRegisters { flag, mut of, mut err, pc, .. }, _) = split_registers(&mut self.registers);
 
                     // LHS argument is always indirect, load it
-                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b as usize)?);
+                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b)?);
 
                     // RHS is only indirect if the flag is set
                     let rhs: $t = if args.indirect_rhs {
-                        $t::from_be_bytes(self.memory.read_bytes(c as usize)?)
+                        $t::from_be_bytes(self.memory.read_bytes(c)?)
                     } else {
                         c.into()
                     };
@@ -104,7 +104,7 @@ macro_rules! wideint_ops {
                     *of = overflow as Word;
                     *err = 0;
 
-                    self.memory.write_bytes(owner_regs, dest_addr as usize, &wrapped.to_be_bytes())?;
+                    self.memory.write_bytes(owner_regs, dest_addr, &wrapped.to_be_bytes())?;
 
                     inc_pc(pc)
                 }
@@ -121,13 +121,13 @@ macro_rules! wideint_ops {
 
                     // LHS is only indirect if the flag is set
                     let lhs: $t = if args.indirect_lhs {
-                        $t::from_be_bytes(self.memory.read_bytes(b as usize)?)
+                        $t::from_be_bytes(self.memory.read_bytes(b)?)
                     } else {
                         b.into()
                     };
                     // RHS is only indirect if the flag is set
                     let rhs: $t = if args.indirect_rhs {
-                        $t::from_be_bytes(self.memory.read_bytes(c as usize)?)
+                        $t::from_be_bytes(self.memory.read_bytes(c)?)
                     } else {
                         c.into()
                     };
@@ -141,7 +141,7 @@ macro_rules! wideint_ops {
                     *of = overflow as Word;
                     *err = 0;
 
-                    self.memory.write_bytes(owner_regs, dest_addr as usize, &wrapped.to_be_bytes())?;
+                    self.memory.write_bytes(owner_regs, dest_addr, &wrapped.to_be_bytes())?;
 
                     inc_pc(pc)
                 }
@@ -157,11 +157,11 @@ macro_rules! wideint_ops {
                     let (SystemRegisters { flag, mut of, mut err, pc, .. }, _) = split_registers(&mut self.registers);
 
                     // LHS is always indirect
-                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b as usize)?);
+                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b)?);
 
                     // RHS is only indirect if the flag is set
                     let rhs: $t = if args.indirect_rhs {
-                        $t::from_be_bytes(self.memory.read_bytes(c as usize)?)
+                        $t::from_be_bytes(self.memory.read_bytes(c)?)
                     } else {
                         c.into()
                     };
@@ -186,7 +186,7 @@ macro_rules! wideint_ops {
 
                     *of = 0;
 
-                    self.memory.write_bytes(owner_regs, dest_addr as usize, &result.to_be_bytes())?;
+                    self.memory.write_bytes(owner_regs, dest_addr, &result.to_be_bytes())?;
 
                     inc_pc(pc)
                 }
@@ -201,9 +201,9 @@ macro_rules! wideint_ops {
                     let owner_regs = self.ownership_registers();
                     let (SystemRegisters { flag, mut of, mut err, pc, .. }, _) = split_registers(&mut self.registers);
 
-                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b as usize)?);
-                    let rhs: $t = $t::from_be_bytes(self.memory.read_bytes(c as usize)?);
-                    let modulus: $t = $t::from_be_bytes(self.memory.read_bytes(d as usize)?);
+                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b)?);
+                    let rhs: $t = $t::from_be_bytes(self.memory.read_bytes(c)?);
+                    let modulus: $t = $t::from_be_bytes(self.memory.read_bytes(d)?);
 
                     let result: $t = if modulus == 0 {
                         if is_unsafe_math(flag.into()) {
@@ -226,7 +226,7 @@ macro_rules! wideint_ops {
 
                     *of = 0;
 
-                    self.memory.write_bytes(owner_regs, dest_addr as usize, &result.to_be_bytes())?;
+                    self.memory.write_bytes(owner_regs, dest_addr, &result.to_be_bytes())?;
 
                     inc_pc(pc)
                 }
@@ -241,9 +241,9 @@ macro_rules! wideint_ops {
                     let owner_regs = self.ownership_registers();
                     let (SystemRegisters { flag, mut of, mut err, pc, .. }, _) = split_registers(&mut self.registers);
 
-                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b as usize)?);
-                    let rhs: $t = $t::from_be_bytes(self.memory.read_bytes(c as usize)?);
-                    let modulus: $t = $t::from_be_bytes(self.memory.read_bytes(d as usize)?);
+                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b)?);
+                    let rhs: $t = $t::from_be_bytes(self.memory.read_bytes(c)?);
+                    let modulus: $t = $t::from_be_bytes(self.memory.read_bytes(d)?);
 
                     let result: $t = if modulus == 0 {
                         if is_unsafe_math(flag.into()) {
@@ -267,7 +267,7 @@ macro_rules! wideint_ops {
 
                     *of = 0;
 
-                    self.memory.write_bytes(owner_regs, dest_addr as usize, &result.to_be_bytes())?;
+                    self.memory.write_bytes(owner_regs, dest_addr, &result.to_be_bytes())?;
 
                     inc_pc(pc)
                 }
@@ -282,9 +282,9 @@ macro_rules! wideint_ops {
                     let owner_regs = self.ownership_registers();
                     let (SystemRegisters { mut of, mut err, pc, flag, .. }, _) = split_registers(&mut self.registers);
 
-                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b as usize)?);
-                    let rhs: $t = $t::from_be_bytes(self.memory.read_bytes(c as usize)?);
-                    let divider: $t = $t::from_be_bytes(self.memory.read_bytes(d as usize)?);
+                    let lhs: $t = $t::from_be_bytes(self.memory.read_bytes(b)?);
+                    let rhs: $t = $t::from_be_bytes(self.memory.read_bytes(c)?);
+                    let divider: $t = $t::from_be_bytes(self.memory.read_bytes(d)?);
 
                     const S: usize = core::mem::size_of::<$t>();
 
@@ -311,7 +311,7 @@ macro_rules! wideint_ops {
                     *of = overflows as Word;
                     *err = 0;
 
-                    self.memory.write_bytes(owner_regs, dest_addr as usize, &result.to_be_bytes())?;
+                    self.memory.write_bytes(owner_regs, dest_addr, &result.to_be_bytes())?;
 
                     inc_pc(pc)
                 }
